+++ src/features/mod.rs
@@ -1,2 +1,5 @@
 #[cfg(feature = "serde")]
 mod serde;
+
+#[cfg(feature = "zerocopy")]
+mod zerocopy;